ecosystem/use_cases/conversational_team/requirements.txt
pyautogen
llama-cpp-python
llama-cpp-python[server]